# contracts/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CaseEvaluation": {
        "additionalProperties": false,
        "properties": {
          "ablation": {
            "$ref": "#/components/schemas/AblationResult"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "outcome": {
            "$ref": "#/components/schemas/OutcomeSnapshot"
          },
          "outcome_evaluation": {
            "$ref": "#/components/schemas/OutcomeEvaluation"
          },
          "partition": {
            "title": "Partition",
            "type": "string"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          },
          "process_evaluation": {
            "$ref": "#/components/schemas/ProcessEvaluation"
          },
          "schema_version": {
            "const": "case-evaluation.v1",
            "default": "case-evaluation.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "topology": {
            "enum": [
              "B0",
              "B1",
              "B2",
              "B3"
            ],
            "title": "Topology",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "partition",
          "topology",
          "ablation",
          "outcome",
          "process_evaluation",
          "outcome_evaluation",
          "passed"
        ],
        "title": "CaseEvaluation",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SoC Operational Decision Twin",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/decision-cases/{case_id}/evaluation": {
      "get": {
        "operationId": "get_evaluation_api_v1_decision_cases__case_id__evaluation_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseEvaluation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evaluation",
        "tags": [
          "evaluations"
        ]
      }
    },
    "/api/v1/decision-cases/{case_id}/evaluations": {
      "post": {
        "operationId": "create_evaluation_api_v1_decision_cases__case_id__evaluations_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseEvaluation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Evaluation",
        "tags": [
          "evaluations"
        ]
      }
    },
    "/api/v1/decision-cases/{case_id}/evidence": {
      "get": {
        "operationId": "get_evidence_api_v1_decision_cases__case_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Evidence"
                  },
                  "title": "Response Get Evidence Api V1 Decision Cases  Case Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evidence",
        "tags": [
          "decision-cases"
        ]
      }
    },
    "/api/v1/decision-cases/{case_id}/outcome-advances": {
      "post": {
        "operationId": "create_outcome_advance_api_v1_decision_cases__case_id__outcome_advances_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutcomeAdvanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutcomeSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Outcome Advance",
        "tags": [
          "outcomes"
        ]
      }
    },
    "/api/v1/decision-cases/{case_id}/review-runs": {
      "post": {
        "operationId": "create_review_run_api_v1_decision_cases__case_id__review_runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "role_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role Id"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReviewRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Command"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Review Run",
        "tags": [
          "agent-runs"
        ]
      }
    },
    "/api/v1/decision-cases/{case_id}/simulated-decisions": {
      "post": {
        "operationId": "create_simulated_decision_api_v1_decision_cases__case_id__simulated_decisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "review_run_id",
            "required": true,
            "schema": {
              "title": "Review Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AblationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Simulated Decision",
        "tags": [
          "simulated-decisions"
        ]
      }
    },
    "/api/v1/decision-cases/{case_id}/timeline": {
      "get": {
        "operationId": "get_timeline_api_v1_decision_cases__case_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Get Timeline Api V1 Decision Cases  Case Id  Timeline Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Timeline",
        "tags": [
          "decision-cases"
        ]
      }
    },
    "/api/v1/decision-cases/{case_id}/workspace": {
      "get": {
        "operationId": "get_workspace_api_v1_decision_cases__case_id__workspace_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionWorkspaceProjection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workspace",
        "tags": [
          "decision-cases"
        ]
      }
    },
    "/api/v1/dev/fixtures/{case_id}/observable": {
      "get": {
        "operationId": "get_observable_packet_api_v1_dev_fixtures__case_id__observable_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Observable Packet Api V1 Dev Fixtures  Case Id  Observable Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Observable Packet",
        "tags": [
          "developer"
        ]
      }
    },
...
  }
}